Updates nav bar and test
