--- client/src/app.css
@@ -1,8 +1,0 @@
-.site-layout .site-layout-background {
-	background: #fff;
-}
-
-.site-layout {
-	padding-top: 70px;
-	margin-top: 84;
-}
