--- client/src/pages/Profile/style.css
@@ -1,33 +1,0 @@
-.bgProfileComponent {
-  background-color: #f6f6f6;
-  box-shadow: 0 17px 18px 0 rgba(0, 0, 0, 0.2),
-    0 9px 20px 0 rgba(0, 0, 0, 0.19);
-}
-
-.custom-style-profile {
-  border-radius: 25px;
-}
-
-.profile-title-style {
-  color: #236184;
-}
-
-.profile-img-style {
-  width: 25%;
-  border-radius: 60%;
-  border: solid #236184;
-}
-
-.btnColor {
-  background-color: #236184;
-}
-
-.text-style {
-  color: #3077a0;
-}
-
-.center {
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-}
